+++ src/main.rs
@@ -7,6 +7,7 @@ mod error;
 mod keybindings;
 mod model;
 mod setup;
+mod sources;
 mod view;
 mod watch;
 mod worker;
@@ -16,10 +17,7 @@ use std::os::fd::IntoRawFd as _;
 
 use std::{
     fmt::Display,
-    fs::read_to_string,
-    io::{self, Read as _, Write as _},
-    path::{Path, PathBuf},
-    process::{Command, Stdio},
+    io::{self, Read as _},
     sync::{
         OnceLock,
         mpsc::{self},
@@ -39,7 +37,6 @@ use ratatui::{
 
 use mdfrier::MarkdownLink;
 use ratatui_image::{picker::ProtocolType, protocol::Protocol, sliced::SlicedProtocol};
-use reqwest::header::CONTENT_TYPE;
 use setup::{SetupResult, setup_graphics};
 
 use crate::{
@@ -48,14 +45,15 @@ use crate::{
     error::Error,
     keybindings::PollResult,
     model::{DocumentId, Model},
+    sources::open_source,
     view::view,
     watch::watch,
     worker::{ImageCache, worker_thread},
 };
 
-const OK_END: &str = " ok.";
+pub const OK_END: &str = " ok.";
 
-static VERSION: OnceLock<String> = OnceLock::new();
+pub static VERSION: OnceLock<String> = OnceLock::new();
 
 fn main() -> io::Result<()> {
     let mut cmd = command!() // requires `cargo` feature
@@ -387,109 +385,6 @@ fn run(terminal: &mut DefaultTerminal, mut model: Model) -> Result<(), Error> {
             })?;
         }
     }
-}
-
-fn open_source(
-    source: &str,
-    url_transform_command: Option<String>,
-) -> Result<(String, Option<PathBuf>, Option<PathBuf>), Error> {
-    use ghrepo::GHRepo;
-    use url::Url;
-
-    log::debug!("source: {source:?}");
-    use core::str::FromStr as _;
-    if let Some(handle) = source.strip_prefix("github:")
-        && let Ok(repo) = GHRepo::from_str(handle)
-    {
-        // We only want to try github if the user explicitly prefixed with "github:...".
-        // Otherwise a path like "dir/file.md" would be a valid GHRepo and cause a useless request.
-        let owner = repo.owner();
-        let name = repo.name();
-        let client = reqwest::blocking::Client::builder()
-            .user_agent(format!(
-                "mdfried/{}",
-                VERSION.get().unwrap_or(&"unknown".to_owned())
-            ))
-            .build()?;
-        for branch in ["master", "main"] {
-            let url = format!(
-                "https://raw.githubusercontent.com/{owner}/{name}/refs/heads/{branch}/README.md"
-            );
-            log::info!("trying github URL: {url}");
-            print!("Fetching URL {url}...");
-            let response = client.get(&url).send()?;
-            if response.status().is_success() {
-                println!("{OK_END}");
-                return Ok((response.text()?, None, None));
-            } else {
-                println!("error.");
-            }
-        }
-        return Err(Error::Io(io::Error::other(format!(
-            "failed to request https://raw.githubusercontent.com/{owner}/{name}/refs/heads/[master|main]/README.md"
-        ))));
-    } else if let Ok(url) = Url::parse(source) {
-        log::info!("requesting URL: {url}");
-        print!("Fetching URL {url}...");
-        let client = reqwest::blocking::Client::builder()
-            .user_agent(format!(
-                "mdfried/{}",
-                VERSION.get().unwrap_or(&"unknown".to_owned())
-            ))
-            .build()?;
-        let response = client.get(url.as_ref()).send()?;
-        if response.status().is_success() {
-            println!("{OK_END}");
-            log::debug!(
-                "have url_transform_command? {}, content_type: {:?}",
-                url_transform_command.is_some(),
-                response.headers().get(CONTENT_TYPE)
-            );
-            if let Some(url_transform_command) = url_transform_command
-                && let Some(content_type) = response.headers().get(CONTENT_TYPE)
-                && content_type
-                    .to_str()
-                    .map_err(|err| {
-                        Error::Io(io::Error::other(format!("content-type error: {err}")))
-                    })?
-                    .starts_with("text/html")
-            {
-                let mut child = Command::new("sh")
-                    .arg("-c")
-                    .arg(url_transform_command)
-                    .env("URL", url.as_ref())
-                    .stdin(Stdio::piped())
-                    .stdout(Stdio::piped())
-                    .spawn()?;
-
-                let Some(stdin) = child.stdin.as_mut() else {
-                    return Err(Error::Io(io::Error::other(
-                        "url_transform_command pipe error",
-                    )));
-                };
-                stdin.write_all(response.text()?.as_bytes())?;
-
-                let output = child.wait_with_output()?;
-
-                return Ok((
-                    String::from_utf8(output.stdout)
-                        .map_err(|_err| Error::Io(io::Error::other("response not utf-8")))?,
-                    None,
-                    None,
-                ));
-            }
-            return Ok((response.text()?, None, None));
-        } else {
-            println!("error.");
-            return Err(Error::Io(io::Error::other(format!(
-                "failed to request {url}"
-            ))));
-        }
-    }
-
-    let path = PathBuf::from(source);
-    let basepath = path.parent().map(Path::to_path_buf);
-    Ok((read_to_string(&path)?, Some(path), basepath))
 }
 
 #[cfg(test)]
